+++ previous-versions/sync-for-commerce-version-1/pkg/models/shared/accountingaccount.go
@@ -3,7 +3,8 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1/pkg/utils"
+	"github.com/ericlagergren/decimal"
 )
 
 type AccountingAccountMetadata struct {
@@ -117,7 +118,7 @@ type AccountingAccount struct {
 	// There are only a very small number of edge cases where this currency code is returned by the Codat system.
 	Currency *string `json:"currency,omitempty"`
 	// Current balance in the account.
-	CurrentBalance *types.Decimal `json:"currentBalance,omitempty"`
+	CurrentBalance *decimal.Big `decimal:"number" json:"currentBalance,omitempty"`
 	// Description for the account.
 	Description *string `json:"description,omitempty"`
 	// Full category of the account.
@@ -148,6 +149,17 @@ type AccountingAccount struct {
 	ValidDatatypeLinks []AccountingAccountValidDataTypeLinks `json:"validDatatypeLinks,omitempty"`
 }
 
+func (a AccountingAccount) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(a, "", false)
+}
+
+func (a *AccountingAccount) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &a, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
 func (o *AccountingAccount) GetCurrency() *string {
 	if o == nil {
 		return nil
@@ -155,7 +167,7 @@ func (o *AccountingAccount) GetCurrency() *string {
 	return o.Currency
 }
 
-func (o *AccountingAccount) GetCurrentBalance() *types.Decimal {
+func (o *AccountingAccount) GetCurrentBalance() *decimal.Big {
 	if o == nil {
 		return nil
 	}